requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
audioread==3.0.1
av==14.3.0
babel==2.17.0
bitsandbytes==0.45.5
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
csvw==3.5.1
ctranslate2==4.6.0
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
dlinfo==2.0.0
dnspython==2.7.0
einops==0.8.1
einops-exts==0.0.4
exceptiongroup==1.3.0
faster-whisper==1.1.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.3.2
greenlet==3.2.2
h11==0.16.0
hf-xet==1.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.1
humanfriendly==10.0
idna==3.10
isodate==0.7.2
Jinja2==3.1.6
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
langchain==0.3.25
langchain-community==0.3.23
langchain-core==0.3.59
langchain-experimental==0.3.4
langchain-huggingface==0.2.0
langchain-mongodb==0.6.1
langchain-ollama==0.3.2
langchain-text-splitters==0.3.8
langchainhub==0.1.21
langsmith==0.3.42
language-tags==1.2.0
lark==1.2.2
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
munch==4.0.0
mypy_extensions==1.1.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.2
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
ollama==0.4.8
onnxruntime==1.22.0
orjson==3.10.18
packaging==24.2
phonemizer==3.3.0
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydub==0.25.1
pymongo==4.12.1
pyparsing==3.2.3
pypdf==5.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==1.5.0
rpds-py==0.25.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
segments==2.3.0
sentence-transformers==4.1.0
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
SQLAlchemy==2.0.40
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
torchaudio==2.7.0
tqdm==4.67.1
transformers==4.51.3
triton==3.3.0
types-requests==2.32.0.20250328
typing==3.7.4.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
yarl==1.20.0
zstandard==0.23.0
